+++ backend/src/main/java/com/diya/backend/dto/dashboard/DashboardKpiDTO.java
@@ -2,10 +2,12 @@ package com.diya.backend.dto.dashboard;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter @Setter @AllArgsConstructor @NoArgsConstructor @Builder
 public class DashboardKpiDTO {
     private int newOrdersToday;
-    private double paymentsReceivedToday;
+    private BigDecimal paymentsReceivedToday;
     private int pendingOrders;
-    private double totalOutstanding;
+    private BigDecimal totalOutstanding;
 }
